requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
elevenlabs==0.2.27
pydub==0.25.1
werkzeug==3.0.1
gunicorn==21.2.0
